+++ appinfo/application.php
@@ -21,9 +21,6 @@ use \OCA\FbSync\Controller\PageController;
 use \OCA\FbSync\Controller\FacebookController;
 use \OCP\IRequest;
 
-
-Application::$l10n = \OC::$server->getL10N('fbsync');
-
 class Application extends App {
 	
 	 /**
@@ -31,11 +28,11 @@ class Application extends App {
 	 * @var array
 	 */
 	static $appname = 'fbsync';
-	static $l10n;
+	public $user;
+	
 	// No PHOTO in $ContactsProbTable
 	public static $index_properties = array('BDAY', 'UID', 'N', 'FN', 'TITLE', 'ROLE', 'NOTE', 'NICKNAME', 'ORG',
 											'CATEGORIES', 'EMAIL', 'TEL', 'IMPP', 'ADR', 'URL', 'GEO', 'CLOUD', 'FBID');
-
 	// From contacts+
 	static $ContactsTable;
 	static $AddrBookTable;
@@ -46,10 +43,17 @@ class Application extends App {
 	const MAXPICTURESIZE = 800;
 	const JAROWINKLERMAX = 85; // Percent for Jaro-Winkler match tolerance
 	
-	public function __construct (array $urlParams=array()) {
+	public function __construct ($user=false, array $urlParams=array()) {
 		
 		parent::__construct(self::$appname, $urlParams);
         $container = $this->getContainer();
+		if(!$user) {
+			$this->user = \OCP\User::getUser();
+		} else {
+			$this->user = $user;
+		}
+		\OCP\Util::writeLog('fbsync', $this->user, \OCP\Util::INFO);
+		
 		
 		// Contact+ compatibility
 		if(\OCP\App::isEnabled('contactsplus')) {
@@ -71,8 +75,8 @@ class Application extends App {
 		 * User home folder
 		 */
 		$container -> registerService('userHome', function(IContainer $c) {
-			$server = $c->query('ServerContainer');
-			return $server->getUserSession()->getUser()->getHome();
+			$datadir = \OCP\Config::getSystemValue('datadirectory');
+			return $datadir.'/'.$this->user;
 			
 		});
 		
@@ -99,19 +103,12 @@ class Application extends App {
 				$c->query('FacebookController')
 			);
 		});
-		
-		/**
-		 * L10N
-		 */
-		$container -> registerService('L10N', function(IContainer $c) {
-			return $c -> query('ServerContainer') -> getL10N($c -> query('AppName'));
-		});
 
 
 		/**
 		 * Contacts
 		 */
-		$container->registerService('Contacts', function($c) {
+		$container->registerService('Contacts', function(IContainer $c) {
 			return new Contacts(
 				$c->query('OCP\Contacts\IManager'),
 				$c->query('OCP\ICache'),