+++ js/sync.js
@@ -10,9 +10,9 @@
 
 var localcontacts = new Array();
 
-function isDoneSyncing(total) {
+function isDoneSyncing(total, syncbutton) {
 	if(total==localcontacts.length) {
-		$("#syncall").text('Done !').removeClass('loading');
+		$(syncbutton).text('Done !').removeClass('loading');
 		$('#loader').remove();
 		$('#contacts-list').fadeIn();
 	}
@@ -31,11 +31,9 @@ function syncPictures() {
 				'apps/fbsync/getphoto/{id}/100',
 				{id: response['id']}
 			);
-			var syncstatus = synced+" contact"+(synced>1?'s':'')+" synced ("+error+" error"+(error>1?'s':'')+")";
 			var contactdivStart = '<div class="sync-contact tooltipped" title="'+response['name'];
 			var contactdivError = ': '+response['error']
 			var contactdivEnd = '"><img src="'+url+'" height="100" width="100" /></div>'
-			$('#syncstatus').text(syncstatus);
 
 			// Error or success?
 			if(response['error']!=false) {
@@ -46,10 +44,67 @@ function syncPictures() {
 				$('#syncsuccess').append(contactdivStart+contactdivEnd);
 			}
 			$('.tooltipped').tipsy();
-			isDoneSyncing(synced+error);
+			isDoneSyncing(synced+error, "#syncall");
+		
+			// Update status
+			var syncstatus = synced+" contact"+(synced>1?'s':'')+" synced ("+error+" error"+(error>1?'s':'')+")";
+			$('#syncstatus').text(syncstatus);
+			
 		}).fail(function() {
+			
 			error++;
-			$('#syncstatus').text(synced+" contact"+(synced>1?'s':'')+" synced ("+error+" error"+(error>1?'s':'')+")");
+			
+			// Update status
+			var syncstatus = synced+" contact"+(synced>1?'s':'')+" synced ("+error+" error"+(error>1?'s':'')+")";
+			$('#syncstatus').text(syncstatus);
+			
+		});
+	})
+}
+
+function syncBirthdays() {
+	
+	$('#loader').fadeIn();
+	
+	var synced = 0;
+	var error = 0;
+	var ignored = 0;
+	localcontacts.forEach(function(id, index, array){
+		$.get(OC.generateUrl('apps/fbsync/setbday/'+id)).done(function(response) {
+
+			var url = OC.generateUrl(
+				'apps/fbsync/getphoto/{id}/100',
+				{id: response['id']}
+			);
+			var contactdivStart = '<div class="sync-contact tooltipped" title="'+response['name']+': '+response['birthday'];
+			var contactdivError = ': '+response['error']
+			var contactdivEnd = '"><img src="'+url+'" height="100" width="100" /></div>'
+			
+			// Error or success?
+			if(response['birthday'] == true) {
+				ignored++;
+			} else if(response['error']!=false) {
+				error++;
+				$('#syncerror').append(contactdivStart+contactdivError+contactdivEnd);
+			} else {
+				synced++;
+				$('#syncsuccess').append(contactdivStart+contactdivEnd);
+			}
+			$('.tooltipped').tipsy();
+			isDoneSyncing(synced+error+ignored, "#syncbday");
+			
+			// Update status
+			var syncstatus = synced+" contact"+(synced>1?'s':'')+" synced ("+error+" error"+(error>1?'s':'')+" & "+ignored+" ignored)";
+			$('#syncstatus').text(syncstatus);
+			
+		}).fail(function() {
+			
+			error++;
+			
+			// Update status
+			var syncstatus = synced+" contact"+(synced>1?'s':'')+" synced ("+error+" error"+(error>1?'s':'')+" & "+ignored+" ignored)";
+			$('#syncstatus').text(syncstatus);
+			
 		});
 	})
 }
@@ -61,7 +116,8 @@ function syncPictures() {
 		
 		// Better visual
 		$('#contacts-list').fadeOut();
-		$('.tooltipped').tipsy();
+		$('.tooltipped-top').tipsy({gravity: 's'});
+		$('.tooltipped-bottom').tipsy({gravity: 'n'});
 		
 		// Hack to resize auto #controls (/core/js/js.js:1435)
 		if($('#controls').length) {
@@ -103,8 +159,14 @@ function syncPictures() {
 //----------  BUTTONS ----------	
 		// Toggle matched button
 		$("#syncall").click(function() {
-			$("#syncall").text('Loading...').prop('disabled',true).addClass('loading');
+			$("#syncall").text('Loading...').addClass('loading');
+			$(".syncbutton").prop('disabled',true);
 			syncPictures();
+		})
+		$("#syncbday").click(function() {
+			$("#syncbday").text('Loading...').addClass('loading');
+			$(".syncbutton").prop('disabled',true);
+			syncBirthdays();
 		})
 		
 	});