+++ templates/error.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * ownCloud - fbsync
+ *
+ * This file is licensed under the Affero General Public License version 3 or
+ * later. See the COPYING file.
+ *
+ * @author NOIJN <fremulon@protonmail.com>
+ * @copyright NOIJN 2015
+ */
+
+script('fbsync', 'login');
+style('fbsync', 'fbsync');
+?>
+
+<div id="app">
+	<div id="app-navigation">
+		<?php print_unescaped($this->inc('part.navigation')); ?>
+		<?php print_unescaped($this->inc('part.settings')); ?>
+	</div>
+
+	<div id="app-content">
+		<div id="loader">
+			<div class="spinner">
+				<div class="rect1"></div>
+				<div class="rect2"></div>
+				<div class="rect3"></div>
+				<div class="rect4"></div>
+				<div class="rect5"></div>
+			</div>
+			<div id="loading-status"><?=$_['error'] ?></div>
+		</div>
+	</div>
+</div>